Temporary fix for powa-dev compose image.
Recent version of postgres image changed the authentication mecanism, which
prevented the powa-archivist images to start.  For now, simply use the same
authentication as previous postgres image behavior, which is to allow all
connection without authentication.  This is obviously super bad for security
reason, but this file isn't meant for production.